move RequestResponse to its own submodule
this should be only used by xknx.io.Tunnel and tests
