+++ test/io_tests/disconnect_test.py
@@ -4,7 +4,8 @@ import unittest
 from unittest.mock import patch
 
 from xknx import XKNX
-from xknx.io import Disconnect, UDPClient
+from xknx.io import UDPClient
+from xknx.io.request_response import Disconnect
 from xknx.knxip import (
     HPAI,
     DisconnectRequest,
