+++ test/io_tests/request_response_test.py
@@ -4,7 +4,8 @@ import unittest
 from unittest.mock import MagicMock, patch
 
 from xknx import XKNX
-from xknx.io import RequestResponse, UDPClient
+from xknx.io import UDPClient
+from xknx.io.request_response import RequestResponse
 from xknx.knxip import DisconnectResponse, KNXIPBody
 
 
@@ -39,7 +40,9 @@ class TestConnectResponse(unittest.TestCase):
             self.loop.run_until_complete(request_response.start())
 
     @patch("logging.Logger.debug")
-    @patch("xknx.io.RequestResponse.send_request", new_callable=AsyncMock)
+    @patch(
+        "xknx.io.request_response.RequestResponse.send_request", new_callable=AsyncMock
+    )
     def test_request_response_timeout(self, _send_request_mock, logger_debug_mock):
         """Test RequestResponse: timeout. No callback shall be left."""
         xknx = XKNX()
@@ -58,7 +61,9 @@ class TestConnectResponse(unittest.TestCase):
         # Callback was removed again
         self.assertEqual(udp_client.callbacks, [])
 
-    @patch("xknx.io.RequestResponse.send_request", new_callable=AsyncMock)
+    @patch(
+        "xknx.io.request_response.RequestResponse.send_request", new_callable=AsyncMock
+    )
     def test_request_response_cancelled(self, _send_request_mock):
         """Test RequestResponse: task cancelled. No callback shall be left."""
         xknx = XKNX()
