+++ test/io_tests/tunnelling_test.py
@@ -5,7 +5,8 @@ from unittest.mock import patch
 
 from xknx import XKNX
 from xknx.dpt import DPTArray
-from xknx.io import Tunnelling, UDPClient
+from xknx.io import UDPClient
+from xknx.io.request_response import Tunnelling
 from xknx.knxip import (
     ErrorCode,
     KNXIPFrame,
