+++ xknx/io/__init__.py
@@ -4,19 +4,14 @@ This package contains all objects managing Tunneling and Routing Connections..
 - KNXIPInterface is the overall managing class.
 - GatewayScanner searches for available KNX/IP devices in the local network.
 - Routing uses UDP/Multicast to communicate with KNX/IP device.
-- Tunnelling uses UDP packets and builds a static tunnel with KNX/IP device.
+- Tunnel uses UDP packets and builds a static tunnel with KNX/IP device.
 """
 # flake8: noqa
-from .connect import Connect
-from .connectionstate import ConnectionState
 from .const import DEFAULT_MCAST_GRP, DEFAULT_MCAST_PORT
-from .disconnect import Disconnect
 from .gateway_scanner import GatewayScanFilter, GatewayScanner
 from .knxip_interface import ConnectionConfig, ConnectionType, KNXIPInterface
-from .request_response import RequestResponse
 from .routing import Routing
 from .tunnel import Tunnel
-from .tunnelling import Tunnelling
 from .udp_client import UDPClient
 
 __all__ = [
