+++ xknx/io/request_response/tunnelling.py
@@ -12,10 +12,9 @@ from xknx.knxip import (
 from .request_response import RequestResponse
 
 if TYPE_CHECKING:
+    from xknx.io.udp_client import UDPClient
     from xknx.telegram import IndividualAddress, Telegram
     from xknx.xknx import XKNX
-
-    from .udp_client import UDPClient
 
 
 class Tunnelling(RequestResponse):
